address issues found with clang-tidy
