+++ Source/Particles/Sorting/SortingUtils.H
@@ -19,18 +19,7 @@
  *
  * \param[inout] v Vector of integers, to be filled by this routine
  */
-void fillWithConsecutiveIntegers( amrex::Gpu::DeviceVector<long>& v )
-{
-#ifdef AMREX_USE_GPU
-    // On GPU: Use amrex
-    auto data = v.data();
-    auto N = v.size();
-    AMREX_FOR_1D( N, i, {data[i] = i;});
-#else
-    // On CPU: Use std library
-    std::iota( v.begin(), v.end(), 0L );
-#endif
-}
+void fillWithConsecutiveIntegers( amrex::Gpu::DeviceVector<long>& v );
 
 /** \brief Find the indices that would reorder the elements of `predicate`
  * so that the elements with non-zero value precede the other elements
